Just a small memorandum taken at last Friday's meeting.
I forgot to upload this, but it’s basically just a small memo I took
down during the events that took place in last Friday’s 3-5 o’clock
class. It’s very brief but it was just intended to jot down how we were
going about determining our course of action for website testing.
Basically, it was just recording what we decided and what actions we
took when deciding. Not sure if it’ll be useful down the line but I
thought it’d be nice to have a record.